Use uint32_t instead of size_t for data format relevant data in water index.

+++ libosmscout/include/osmscout/GenWaterIndex.h
@@ -100,12 +100,12 @@ namespace osmscout {
   private:
     double           cellWidth;
     double           cellHeight;
-    size_t           cellXStart;
-    size_t           cellXEnd;
-    size_t           cellYStart;
-    size_t           cellYEnd;
-    size_t           cellXCount;
-    size_t           cellYCount;
+    uint32_t         cellXStart;
+    uint32_t         cellXEnd;
+    uint32_t         cellYStart;
+    uint32_t         cellYEnd;
+    uint32_t         cellXCount;
+    uint32_t         cellYCount;
 
     double           minLon;
     double           minLat;
